requirements.txt
fastapi
uvicorn[standard]
matplotlib
seaborn
python-dotenv
Pillow>=9.0.0